% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{content_box}
\alias{content_box}
\title{Group related content in dashboard}
\usage{
content_box(title, ...)
}
\arguments{
\item{title}{Title string}

\item{...}{Content elements}
}
\value{
HTML defining the content
}
\description{
Group related content in dashboard
}
\examples{
\dontrun{
content_box(
  "I have content!",
  shiny::selectInput("input", "Choose", c("Option 1", "Option 2")),
  shiny::uiOutput("output")
)
}
}
